# docs/technical/api/openapi.yaml
openapi: 3.0.3
info:
  title: CLARIHR APIs
  version: 0.0-bootstrap
  description: |
    Partial hand-maintained OpenAPI supplement for externally relevant core and backoffice
    endpoints that need a checked-in contract before automated export/versioning is introduced.

    The exhaustive contract is still generated at runtime by Swagger in Development
    at /swagger/v1/swagger.json.

    Public API contracts expose `publicId` or `<entity>PublicId`; internal `id` and
    `internalId` never leave the backend. When a resource has business codes, the
    public contract exposes `code` and `normalizedCode`, both normalized to uppercase.

    Core API endpoints require tokens emitted with `client_type=core`. Platform
    backoffice endpoints under `/api/platform/*` require tokens emitted with
    `client_type=platform`.

    All `/api` routes emit `Cache-Control: no-store`, `Pragma: no-cache`, and
    `Expires: 0` to avoid caching sensitive HR or platform responses.
servers:
  - url: /
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PersonnelFileDocumentMetadataResponse:
      type: object
      required:
        - id
        - documentType
        - fileName
        - contentType
        - sizeBytes
        - isActive
        - concurrencyToken
        - createdAtUtc
      properties:
        id:
          type: string
          format: uuid
        documentType:
          type: string
        observations:
          type: string
          nullable: true
        deliveryDate:
          type: string
          format: date-time
          nullable: true
        loanDate:
          type: string
          format: date-time
          nullable: true
        returnDate:
          type: string
          format: date-time
          nullable: true
        fileName:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
          format: int32
        isActive:
          type: boolean
        concurrencyToken:
          type: string
          format: uuid
        createdAtUtc:
          type: string
          format: date-time
        modifiedAtUtc:
          type: string
          format: date-time
          nullable: true
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        traceId:
          type: string
          nullable: true
          description: Correlation identifier that links the error response with server logs.
        errors:
          type: object
          nullable: true
          description: Field-level validation messages. Present when `code` is `common.validation`.
          additionalProperties:
            type: array
            items:
              type: string
        details:
          type: array
          nullable: true
          description: Optional structured metadata returned by some authorization or policy failures.
          items:
            $ref: '#/components/schemas/ProblemDetailItem'
    ProblemDetailItem:
      type: object
      properties:
        resourceKey:
          type: string
          nullable: true
        action:
          type: string
          nullable: true
        fieldKey:
          type: string
          nullable: true
        endpoint:
          type: string
          nullable: true
...
paths:
...
  /api/v1/personnel-files/{id}/documents:
    get:
      tags:
        - personnel-files
      operationId: GetPersonnelFileDocuments
      summary: List lightweight personnel file document previews
      description: Returns document metadata for a personnel file without streaming the binary payload. Use this route to build attachment lists and call the download endpoint only for the selected document.
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document metadata returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonnelFileDocumentMetadataResponse'
        '401':
          description: Authentication is required.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The authenticated user does not have permission to read personnel files or the tenant scope does not match.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: The target personnel file was not found in the tenant scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - personnel-files
      operationId: UploadPersonnelFileDocument
      summary: Upload a personnel file document
      description: Uploads a new personnel file document using multipart form data and returns the persisted metadata without the binary payload. Files are rejected before full buffering when they exceed 10 MiB or when extension, declared MIME type, or basic file signature is not allowed. Allowed formats are PDF, JPG/JPEG, PNG, and DOCX.
      security:
        - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - documentType
                - file
                - concurrencyToken
              properties:
                documentType:
                  type: string
                observations:
                  type: string
                  nullable: true
                deliveryDate:
                  type: string
                  format: date-time
                  nullable: true
                loanDate:
                  type: string
                  format: date-time
                  nullable: true
                returnDate:
                  type: string
                  format: date-time
                  nullable: true
                concurrencyToken:
                  type: string
                  format: uuid
                file:
                  type: string
                  format: binary
                  description: Binary document payload. Maximum 10 MiB. Allowed formats are .pdf, .jpg, .jpeg, .png and .docx.
      responses:
        '201':
          description: Document uploaded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonnelFileDocumentMetadataResponse'
        '400':
          description: Validation error, missing file, unsupported content type or malformed payload.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '413':
          description: The uploaded document exceeds the 10 MiB size limit.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Authentication is required.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The authenticated user does not have permission to manage personnel files or the tenant scope does not match.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: The target personnel file was not found in the tenant scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Business conflict such as stale concurrency token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Business-rule violation, such as invalid document dates.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/personnel-files/documents/{documentId}/inactivate:
    patch:
      tags:
        - personnel-files
      operationId: InactivatePersonnelFileDocument
      summary: Inactivate a personnel file document
      description: Performs a logical inactivation of the selected personnel file document using the document concurrency token.
      security:
        - Bearer: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConcurrencyRequest'
      responses:
        '200':
          description: Document inactivated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonnelFileDocumentMetadataResponse'
        '400':
          description: Validation error, such as missing concurrency token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Authentication is required.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The authenticated user does not have permission to manage personnel files or the tenant scope does not match.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: The target document was not found in the tenant scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Business conflict such as stale concurrency token.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/personnel-files/documents/{documentId}/download:
    get:
      tags:
        - personnel-files
      operationId: DownloadPersonnelFileDocument
      summary: Download a personnel file document
      description: Streams the binary payload for a previously uploaded personnel file document.
      security:
        - Bearer: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document downloaded successfully.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Authentication is required.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: The authenticated user does not have permission to read personnel files or the tenant scope does not match.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: The target document was not found in the tenant scope.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...